requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pyjwt>=2.10.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
openai>=1.60.0
anthropic>=0.40.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
litellm>=1.55.0
